+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 01.01.2019.
 
-- TODO: Describe initial release.
+## [0.1.1] - 03.03.2019
 
-## [0.1.1] - added the possibility to hide the selected check animation
+- added the possibility to hide the selected check animation
+
+## [0.1.2] - 20.09.2019
+
+- added canMultiSelect
+- fixed itemwidth bug with subitems